// src/app/models/client.ts
export interface Client{
    id: number;
    fechaActualizacion: string;
    provincia: string;
    municipio: string;
    nombre: string;
    primerApellido: string;
    segundoApellido: string;
    telefono: string;
    email: string;
    direccion: string;
    codigoPostal: string;
    nombreProducto: string;
    tipoProducto: string;
    imageProducto: string;
}